:wrench: (home): refactor switch and constan
