+++ src/screens/home/RenderCards.tsx
@@ -1,46 +1,17 @@
+import { ReactNode } from 'react'
 import { LinkIcon } from '@chakra-ui/icons'
 import { Button, Flex, Heading } from '@chakra-ui/react'
-import Image from 'next/image'
 import Link from 'next/link'
-
-import {
-  ApolloLogo,
-  ChakraUILogo,
-  NextJsLogo,
-  ReactQueryLogo,
-} from 'assets/svg'
 
 import { STACK } from './constant'
 
 type StackItem = {
   label: string
   url: string
-  logo: number
+  logo: ReactNode
 }
 
 export const RenderCards = () => {
-  const renderLogo = (logo: number) => {
-    switch (logo) {
-      case 0:
-        return <ChakraUILogo />
-      case 1:
-        return <NextJsLogo />
-      case 2:
-        return <ApolloLogo />
-      case 3:
-        return (
-          <Image
-            src={'/NextAuth.png'}
-            alt="Picture of the author"
-            width={80}
-            height={80}
-          />
-        )
-      default:
-        return <ReactQueryLogo />
-    }
-  }
-
   return (
     <Flex justifyContent="center" flexWrap="wrap" gridGap="20px" maxH="60vh">
       {STACK.map((item: StackItem, index: number) => {
@@ -59,7 +30,7 @@ export const RenderCards = () => {
             padding="61px 38px"
           >
             <Flex flex="6" w="100%" justifyContent="center" p="10px">
-              {renderLogo(item.logo)}
+              {item.logo}
             </Flex>
             <Flex flex="10" w="100%" justifyContent="center" p="10px">
               <Heading m="0" color="white" textAlign="center">
