--- src/screens/home/constant.ts
@@ -1,11 +1,0 @@
-export const STACK = [
-  { label: 'Chakra UI', url: 'https://chakra-ui.com/', logo: 0 },
-  { label: 'NextJS', url: 'https://nextjs.org/', logo: 1 },
-  { label: 'Apollo Graphql', url: 'https://www.apollographql.com/', logo: 2 },
-  { label: 'NextAuth', url: 'https://next-auth.js.org/', logo: 3 },
-  {
-    label: 'React Query',
-    url: 'https://react-query-v3.tanstack.com/',
-    logo: 4,
-  },
-]
